Change observe value(from `Boolean` to `Response<T>`), if `LoginSuccess`, save `JwtToken` in `SharedPreferences`
